Action Menus

+++ airy/extensions/roles/buttons/menu.py
@@ -61,14 +61,17 @@ class RemoveModal(miru.Modal):
 
 class MenuView(MenuViewAuthorOnly):
     def __init__(self, ctx: AirySlashContext, channel_id: hikari.Snowflake, message_id: hikari.Snowflake):
+        self.ctx = ctx
+        self.model: t.Optional[ActionMenusModel] = None
+
         self.channel_id = channel_id
         self.message_id = message_id
+
         self.acm_message: t.Optional[hikari.Message] = None
         self.acm_embed: t.Optional[hikari.Embed] = None
+        self.acm_view: t.Optional[miru.View] = None
 
-        self.ctx = ctx
         super().__init__(ctx)
-        self.model: t.Optional[ActionMenusModel] = None
         for item in self.default_buttons:
             self.add_item(item)
 
@@ -78,7 +81,7 @@ class MenuView(MenuViewAuthorOnly):
 
     @property
     def default_buttons(self):
-        return [AddButtonButton(), RemoveButtonButton(), DestroyButton(), PreviewButton(), QuitButton()]
+        return [AddButtonButton(), RemoveButtonButton(), DestroyButton(), PreviewButton(), QuitButton(), SelectEmbed()]
 
     def get_kwargs(self):
         embed = hikari.Embed(title="Action Menus",
@@ -87,9 +90,9 @@ class MenuView(MenuViewAuthorOnly):
         entries_description = []
 
         for index, entry in enumerate(self.model.buttons, 1):
-            entry_role = self.ctx.bot.cache.get_role(hikari.Snowflake(entry.payload))
-            entries_description.append(f"**{index}.** {entry.label} - {entry_role.mention} "
-                                       f"(Button ID: {entry.id}, Role ID: {entry_role.id})")
+            role = self.ctx.bot.cache.get_role(hikari.Snowflake(entry.payload))
+            entries_description.append(f"**{index}.** {entry.label if entry.label else 'Missing'} - {role.mention}"
+                                       f"(Role ID: {role.id})")
 
         embed.description = '\n'.join(entries_description)
         return dict(embed=embed, components=self.build(), flags=self.flags)
@@ -102,6 +105,7 @@ class MenuView(MenuViewAuthorOnly):
         try:
             self.acm_message = await self.ctx.bot.rest.fetch_message(self.channel_id, self.message_id)
             self.acm_embed = self.acm_message.embeds[0]
+            self.acm_view = miru.View.from_message(self.acm_message, timeout=None)
         except hikari.NotFoundError:
             await self.ctx.respond(embed=RespondEmbed.error("Provided action menus missing"))
             return
@@ -119,6 +123,31 @@ class MenuView(MenuViewAuthorOnly):
         message = await self.ctx.interaction.fetch_initial_response()
         super(MenuView, self).start(message)
 
+    async def save(self):
+        try:
+            if len(self.model.buttons.related_objects) == 0:
+                await self.ctx.bot.rest.delete_message(self.channel_id, self.message_id)
+                await self.model.delete()
+                await self.last_ctx.edit_response(embed=RespondEmbed.success("Action Menus was deleted"))
+                self.stop()
+            else:
+                await self.ctx.bot.rest.edit_message(self.channel_id,
+                                                     self.message_id,
+                                                     embed=self.acm_embed,
+                                                     components=self.acm_view.build())
+                for item in self.children:
+                    item.disabled = True
+                kwargs = self.get_kwargs()
+                await self.last_ctx.edit_response(**kwargs)
+                self.stop()
+        except hikari.NotFoundError:
+            pass
+        except hikari.ForbiddenError:
+            embed = RespondEmbed.error(
+                title="Insufficient permissions",
+                description=f"The bot cannot edit message due to insufficient permissions.")
+            await self.last_ctx.respond(embed=embed, flags=hikari.MessageFlag.EPHEMERAL)
+
 
 ViewT = t.TypeVar("ViewT", bound=MenuView)
 
@@ -135,13 +164,12 @@ class AddButtonButton(miru.Button[ViewT]):
 
         if role and str(role.id) not in [entry.payload for entry in self.view.model.buttons]:
             channel_id = self.view.channel_id
-            message_id = self.view.message_id
 
-            entry_model = ActionMenusButtonModel(id_id=self.view.model.id,
+            entry_model = ActionMenusButtonModel(menus_id=self.view.model.id,
                                                  payload=str(role.id),
                                                  style=modal.style,
                                                  action_type=ActionType.ROLE,
-                                                 emoji=modal.emoji.__str__()
+                                                 emoji=modal.emoji
                                                  )
             button = miru.Button(
                 custom_id=f"ACM:{channel_id}:{role.id}",
@@ -149,22 +177,9 @@ class AddButtonButton(miru.Button[ViewT]):
                 label=modal.label,
                 style=modal.style,
             )
-
-            message = await self.view.bot.rest.fetch_message(channel_id, message_id)
-            view = miru.View.from_message(message, timeout=None)
-            view.add_item(button)
-            try:
-                await message.edit(components=view.build())
-            except hikari.NotFoundError:
-                pass
-            except hikari.ForbiddenError:
-                embed = RespondEmbed.error(
-                    title="Insufficient permissions",
-                    description=f"The bot cannot edit message due to insufficient permissions.")
-                await context.respond(embed=embed, flags=hikari.MessageFlag.EPHEMERAL)
-            else:
-                await entry_model.save()
-                self.view.model.buttons.related_objects.append(entry_model)
+            self.view.acm_view.add_item(button)
+            await entry_model.save()
+            self.view.model.buttons.related_objects.append(entry_model)
 
         await self.view.send(modal.get_response_context())
 
@@ -179,38 +194,16 @@ class RemoveButtonButton(miru.Button[ViewT]):
         await modal.wait()
         role = modal.role
         channel_id = self.view.channel_id
-        message_id = self.view.message_id
 
         if role and str(role.id) in [entry.payload for entry in self.view.model.buttons]:
-            if len(self.view.model.buttons.related_objects) == 1:
-                try:
-                    await self.view.ctx.bot.rest.delete_message(channel_id, message_id)
-                except hikari.NotFoundError:
-                    pass
-                except hikari.ForbiddenError:
-                    embed = RespondEmbed.error(
-                        title="Insufficient permissions",
-                        description=f"The bot cannot edit message due to insufficient permissions.")
-                    await context.respond(embed=embed, flags=hikari.MessageFlag.EPHEMERAL)
-                else:
-                    await self.view.model.delete()
-                    await context.edit_response(embed=RespondEmbed.success("Action Menus was deleted"))
-            else:
-                await ActionMenusButtonModel.filter(id_id=self.view.model.id, role_id=role.id).delete()
-                for entry in self.view.model.buttons.related_objects:
-                    if entry.payload == str(role.id):
-                        try:
-                            message = await self.view.bot.rest.fetch_message(channel_id, message_id)
-                        except hikari.NotFoundError:
-                            pass
-                        else:  # Remove button if message still exists
-                            view = miru.View.from_message(message, timeout=None)
+            await ActionMenusButtonModel.filter(menus_id=self.view.model.id, payload=str(role.id)).delete()
 
-                            for item in view.children:
-                                if item.custom_id == f"ACM:{channel_id}:{entry.payload}":
-                                    view.remove_item(item)
-                            await message.edit(components=view.build())
-                        self.view.model.buttons.related_objects.remove(entry)
+            for entry in self.view.model.buttons.related_objects:
+                if entry.payload == str(role.id):
+                    for item in self.view.acm_view.children:
+                        if item.custom_id == f"ACM:{channel_id}:{entry.payload}":
+                            self.view.acm_view.remove_item(item)
+                    self.view.model.buttons.related_objects.remove(entry)
 
         await self.view.send(modal.get_response_context())
 
@@ -226,9 +219,9 @@ class DestroyButton(miru.Button[ViewT]):
             pass
 
         await self.view.model.delete()
-        await context.respond(embed=RespondEmbed.success("Action Menus was deleted"),
-                              components=[],
-                              flags=self.view.flags)
+        await context.edit_response(embed=RespondEmbed.success("Action Menus was deleted"),
+                                    components=[],
+                                    flags=self.view.flags)
         self.view.stop()
 
 
@@ -242,9 +235,20 @@ class PreviewButton(miru.Button[ViewT]):
 
 class QuitButton(miru.Button[ViewT]):
     def __init__(self) -> None:
-        super().__init__(style=hikari.ButtonStyle.PRIMARY, label="Quit", emoji=MenuEmojiEnum.SAVE)
+        super().__init__(style=hikari.ButtonStyle.PRIMARY, label="Save", emoji=MenuEmojiEnum.SAVE)
 
     async def callback(self, context: miru.ViewContext) -> None:
+        try:
+            await self.view.acm_message.edit(embed=self.view.acm_embed, components=self.view.acm_view.build())
+        except hikari.NotFoundError:
+            pass
+        except hikari.ForbiddenError:
+            embed = RespondEmbed.error(
+                title="Insufficient permissions",
+                description=f"The bot cannot edit message due to insufficient permissions.")
+            await context.respond(embed=embed, flags=hikari.MessageFlag.EPHEMERAL)
+            return
+
         for item in self.view.children:
             item.disabled = True
         kwargs = self.view.get_kwargs()
@@ -253,24 +257,25 @@ class QuitButton(miru.Button[ViewT]):
 
 
 class TitleModal(miru.Modal):
-    title_input = miru.TextInput(label="Title",
-                                 placeholder="The title of the embed.", max_length=128)
-
-    def __init__(self, ) -> None:
+    def __init__(self) -> None:
+        self.title_input = miru.TextInput(label="Title",
+                                          placeholder="The title of the embed.", max_length=100,
+                                          style=hikari.TextInputStyle.PARAGRAPH)
         super().__init__("Title")
-        self.title: t.Optional[str] = None
+        self.title_: t.Optional[str] = None
         self.add_item(self.title_input)
 
     async def callback(self, ctx: miru.ModalContext) -> None:
         title = ctx.values.get(self.title_input)
-        self.title = title if title else hikari.UNDEFINED
+        self.title_ = title if title else hikari.UNDEFINED
 
 
 class DescriptionModal(miru.Modal):
-    description_input = miru.TextInput(label="Description",
-                                       placeholder="The description of the embed.", max_length=4096)
-
-    def __init__(self, ) -> None:
+    def __init__(self) -> None:
+        self.description_input = miru.TextInput(label="Description",
+                                                placeholder="The description of the embed.", max_length=4000,
+                                                style=hikari.TextInputStyle.PARAGRAPH
+                                                )
         super().__init__("Description")
         self.description: t.Optional[str] = None
         self.add_item(self.description_input)
@@ -281,10 +286,10 @@ class DescriptionModal(miru.Modal):
 
 
 class ColorModal(miru.Modal):
-    color_input = miru.TextInput(label="Color",
-                                 placeholder="The color of the embed.", min_length=1)
-
     def __init__(self, ) -> None:
+        self.color_input = miru.TextInput(label="Color",
+                                          placeholder="The color of the embed.", min_length=1)
+
         super().__init__("Enter Role")
         self.color: t.Optional[hikari.Color] = None
         self.add_item(self.color_input)
@@ -294,13 +299,16 @@ class ColorModal(miru.Modal):
 
 
 class AuthorModal(miru.Modal):
-    author_input = miru.TextInput(label="Author",
-                                  placeholder="The author of the embed. Appears above the title.", max_length=128)
-
-    author_url_input = miru.TextInput(label="Author URL",
-                                      placeholder="An URL pointing to an image to use for the author's avatar.")
-
     def __init__(self, ) -> None:
+        self.author_input = miru.TextInput(label="Author",
+                                           placeholder="The author of the embed. Appears above the title.",
+                                           style=hikari.TextInputStyle.PARAGRAPH,
+                                           max_length=100)
+
+        self.author_url_input = miru.TextInput(label="Author URL",
+                                               placeholder="An URL pointing to an image to use for the author's avatar.",
+                                               max_length=100)
+
         super().__init__("Enter Role")
         self.author_url: t.Optional[str] = None
         self.author: t.Optional[str] = None
@@ -315,14 +323,15 @@ class AuthorModal(miru.Modal):
 
 
 class FooterModal(miru.Modal):
-    footer_input = miru.TextInput(label="Footer",
-                                  placeholder="The footer of the embed.",
-                                  max_length=256)
-    footer_url_input = miru.TextInput(label="Footer URL",
-                                      placeholder="An url pointing to an image to use for the embed footer.",
-                                      max_length=256)
-
     def __init__(self, ) -> None:
+        self.footer_input = miru.TextInput(label="Footer",
+                                           placeholder="The footer of the embed.",
+                                           max_length=200,
+                                           style=hikari.TextInputStyle.PARAGRAPH)
+        self.footer_url_input = miru.TextInput(label="Footer URL",
+                                               placeholder="An url pointing to an image to use for the embed footer.",
+                                               max_length=200)
+
         super().__init__("Enter Role")
         self.footer: t.Optional[str] = None
         self.footer_url: t.Optional[str] = None
@@ -337,11 +346,11 @@ class FooterModal(miru.Modal):
 
 
 class ThumbnailModal(miru.Modal):
-    thumbnail_input = miru.TextInput(label="Thumbnail URL",
-                                     placeholder="An url pointing to an image to use for the thumbnail.",
-                                     max_length=256)
-
     def __init__(self, ) -> None:
+        self.thumbnail_input = miru.TextInput(label="Thumbnail URL",
+                                              placeholder="An url pointing to an image to use for the thumbnail.",
+                                              max_length=200)
+
         super().__init__("Enter Role")
         self.thumbnail: t.Optional[str] = None
         self.add_item(self.thumbnail_input)
@@ -354,19 +363,20 @@ class ThumbnailModal(miru.Modal):
 
 
 class SelectEmbed(miru.Select[ViewT]):
-    options = [
-        miru.SelectOption(label="Title", value="__title", description="The title of the embed."),
-        miru.SelectOption(label="Description", value="__description", description="The title of the embed."),
-        miru.SelectOption(label="Color", value="__color", description="The color of the embed."),
-        miru.SelectOption(label="Author", value="__author",
-                          description="The author of the embed. Appears above the title."),
-        miru.SelectOption(label="Footer", value="__footer", description="The footer of the embed."),
-        miru.SelectOption(label="Thumbnail", value="__thumbnail",
-                          description="An url pointing to an image to use for the thumbnail.")
-    ]
-
     def __init__(self):
-        super().__init__(options=self.options)
+        self.options = [
+            miru.SelectOption(label="Title", value="__title", description="The title of the embed."),
+            miru.SelectOption(label="Description", value="__description", description="The title of the embed."),
+            miru.SelectOption(label="Color", value="__color", description="The color of the embed."),
+            miru.SelectOption(label="Author", value="__author",
+                              description="The author of the embed. Appears above the title."),
+            miru.SelectOption(label="Footer", value="__footer", description="The footer of the embed."),
+            miru.SelectOption(label="Thumbnail", value="__thumbnail",
+                              description="An url pointing to an image to use for the thumbnail.")
+        ]
+        super().__init__(options=self.options,
+                         placeholder="Embed Settings",
+                         row=2)
 
     async def callback(self, context: miru.ViewContext) -> None:
         value = self.values[0]
@@ -374,7 +384,7 @@ class SelectEmbed(miru.Select[ViewT]):
             modal = TitleModal()
             await context.respond_with_modal(modal)
             await modal.wait()
-            self.view.acm_embed.title = modal.title
+            self.view.acm_embed.title = modal.title_
         elif value == "__description":
             modal = DescriptionModal()
             await context.respond_with_modal(modal)
